feat: 보너스볼 입력 및 변경된 순위 출력하도록 추가

+++ src/main/java/lotto/LottoApplication.java
@@ -7,7 +7,7 @@ import lotto.view.ResultView;
 
 public class LottoApplication {
 	public static void main(String[] args) {
-		int payment = InputView.scanPayment();
+		long payment = InputView.scanPayment();
 		LottoRunner runner = new LottoRunner(payment);
 		ResultView.printLottoNumbers(runner.getLottoNumbers());
